Test/requirements.txt
Django==1.11.3
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
pytz==2017.3
